added a short name for tasks which is the name of the test without spaces.
#44 is completed

put in commands for the future changes in ProjectProperties. It will no longer have a path for template and submissions separated but it will just have 1 path for the project location

Reordered the ordering of the beans in applicationContext.xml to load the properties first.

Modified viewAll/unitTest.jsp to contain a hyperlink to the test.

git-svn-id: http://svn.it.usyd.edu.au/svn/infosubsys@29 26cd2a52-c9c5-0410-8157-a43b28e47565