Changed the color of the green to be a more visually aesthetic green
Created a popup to allow new assessments to be created.

New assessments can be created. Currently the dueDate of the assessment is hard coded. This is due to a bug in the datepicker that is refusing to initialize itself.
Assessments can be created and they are being written to disk as well as stored in memory.

Changed the build.xml to use java 1.7 instead of 1.5.

Modified all output code that used "\r\n" to denote a line ending to use System.getProperty("line.separator"). This is OS independent and will create the correct line ending for the platform the submission system is run on.

Added links in the decorator for easier navigation. This will be limited to only tutors in the future.

git-svn-id: http://svn.it.usyd.edu.au/svn/infosubsys@48 26cd2a52-c9c5-0410-8157-a43b28e47565